Update README.md
Learn More about the DAO desentreliced autonomus organisation AMS-DAO is secuyred for building arcademachines, for use outdoors and suncell, armored iron box. the first donations are tokenized as the smart-contract dictates. living on the poligon blockchain as the poor kids school charity are going to build an organisation with the members owning an NFT a dobbel solution for online identity card and crypto wallet.

the first 27 are paying for the rest, 10000 AMS collectable serie for a good case.the donations are blockchain and you can trace every block. the members are velcom to use the camera in the AMS and watch the poor kids learning. 

the arcademachineschool is an online school. the students can learn for free, every cource with an certificate, the AMS can bve equipped with an printer or make an online certification use it to get an job and earn money. 

the first roadmap and no one believed that we could achieve what we set out to do. 10.000 NFTs on opensea.io/arcademachineschool and rarible and Nifty Gateway. you can read my pdf withepaper and road map, the AMS-DAO GOVERNANCE AND VOTING IS CONTROLLED BY THE MEMBERS. help us help everybody witout school using an online course only. 

help by promoting the AMS project. 57 million reasons for you to bye an NFT. and 57 million reasons for me to build a DAO charity project. the student got cources from 1 week to monts. autotranslate the info for thestudent. read and write is the first chance to make a greate new virtual metaverse. 

Some of my NFT collections have integrated to help build a DAO-NFT charity educational services and new metods to help the students locate an cource and compleate it as fast as possible. the main goale is to help poor kods help themselfes. thought it would take a group of members who are intrested in educate the poor. geting certified and post adds for work. 

We've finaly done it and we’re only getting started. I’d like to thank you for placing your faith and donation in the AMS-DAO charity, education for all. africa and asia, india is the main locations of the 57 millions of poor children. in the loving memmory of my mother Kirsten Anett Gronli you are missed always mamma.

place outdoor education machines, support and repair is the finaly educated for the first job, to do service and repairs.AMS-DAO arcademachineschool NFT collectable serie for bulding an desentreliced educational service. We remain relentless at building better ways for an easy and fast education. join us and get a vote and governance of the assets, got a good carible idea lets build bether educaton AMS-DAO on the poligon blockchain. suport the project and join me provide educational resource for the poorest 57 million children. AMS-DAO NFTs is looking forward to collaborating with you.



Signed-off-by: Ingeborg L Løyning <101161391+GDPOT@users.noreply.github.com>